[REFACTOR] City구조체와 CityInfo구조체 생성 #3
